LayoutTests/discovery/TC-IntermittentServerService.html

<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>

<script>

var devdropped;

description('Tests Acceptable UA behavior when Intermittent Server Service is dicovered.');

function getNetworkServices(type, okCallback, errCallback) {
	devdropped = false;
	navigator.getNetworkServices(type, okCallback, errCallback);
}

var srvs;

function okAddDev(services) {
	testPassed("navigator.getNetworkServices() first call to ok callback.");
	srvs = services;
	shouldBe("srvs.servicesAvailable==0", "true");
	srvs.onserviceavailable = addDevCB;
	srvs.onserviceunavailable = dropDevCB
}

function okCallback2(services) {
	testPassed("navigator.getNetworkServices() second call to ok callback.");
	srvs = services;
	shouldBe("srvs.servicesAvailable>0", "true");
	srvs.onserviceavailable = addDevCB;
	srvs.onserviceunavailable = dropDevCB
}

function errShouldNotbeCalled(err) {
	testFailed('GetNetworkServices() should have called ok Callback.');
}

function dropDevCB() {
	testPassed("Service dropped out.");
	devdropped = true;
}

function addDevCB() {
	testPassed("Service came back.");

	if (devdropped) {
		srvs.onserviceavailable = null;
		srvs.onserviceunavailable = null;
		finishJSTest();
	} else {
		navigator.getNetworkServices(
			"upnp:urn:schemas-upnp-org:service:RemoteUIServer:1", okCallback2, errCallback);
	}
}

getNetworkServices("upnp:urn:schemas-upnp-org:service:RemoteUIServer:1", okAddDev, errShouldNotbeCalled);
window.jsTestIsAsync = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
